Vendor bindings file specifically for docs
Docs.rs does not allow network access, and will not add support for the esp fork
of Rust. Therefore we need to vendor the bindings file in Git so that docs.rs
can build the docs without these things. The build script will not attempt to
run under docs.rs at all, and when it is running a real build the bindings file
will be automatically updated by the build script.