Fix `ESP_IDF_SDKCONFIG` and `_DEFAULTS` containing backslashes on windows.
cmake doesn't allow backslashes in its function arguments, so we convert these paths to paths with slashes.
Currently this also forbids non-unicode paths, because we have to convert the `OsStr` to `str` to do this
replace operation (without us having to implement it ourselves).

This fixes the native build on windows if either `ESP_IDF_SDKCONFIG` and/or
`ESP_IDF_SDKCONFIG_DEFAULTS` contain paths with backslashes.