Prevent setting CMAKE_BUILD_PROFILE
Normally the `cmake` crate sets this variable depending on the current cargo profile,
but we don't want this as it interferes with esp-idf's own configuration.

The esp-idf is configured with kconfig and in that system there already exist options
for configuring optimization settings (using sdkconfig). Additionally, while compiling
the esp-idf, the bootloader is compiled too and setting `CMAKE_BUILD_PROFILE` would
also interfere with bootloader optimization options which are seperate from the application's.