Allow build configuration to be specified in manifest metadata
This change allows the build configuration to be specified in the `package.metadata.esp-idf-sys`
table of the root crate's manifest (Cargo.toml). If no root crate exists in the workspace the crate
with the name equal to the `ESP_IDF_SYS_ROOT_CRATE` environment variable is used.

All configuration specified via environment variables will override the corresponding options
of the cargo metadata. If `cargo metadata` fails for some reason, this will only generate a warning
in the build output and not cause the build to fail.

Add build dependency `cargo_metadata`.