Include the I2S 5.x driver headers.
In ESP-IDF 5.0, the I2S driver underwent a major rewrite. The new
driver requires including a new set of header files. The old driver
is still available (and is still included here and has bindings
provided) but is deprecated and slated for removal.

This is required for the I2S driver being written for esp-idf-hal.